pyproject.toml
[project]
name = "market-insights-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "openai>=1.6.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
